Add Handler class to handle commands
